* Removed ENABLE_FORCE_SCRIPT logic - scripting is here to stay. * Fixed numerous warnings.

+++ TheForceEngine/TFE_ExternalData/dfLogics.cpp
@@ -126,7 +126,7 @@ namespace TFE_ExternalData
 
 		if (cJSON_IsNumber(data) && strcasecmp(data->string, "fieldOfView") == 0)
 		{
-			customLogic.fov = floatToAngle(data->valueint);
+			customLogic.fov = floatToAngle((f32)data->valueint);
 			return true;
 		}
 
@@ -330,7 +330,7 @@ namespace TFE_ExternalData
 
 		if (cJSON_IsNumber(data) && strcasecmp(data->string, "rotationSpeed") == 0)
 		{
-			customLogic.rotationSpeed = floatToAngle(data->valueint);
+			customLogic.rotationSpeed = floatToAngle((f32)data->valueint);
 			return true;
 		}
